PR comments #2
