+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="net.measurementlab.ndt.android">
+    package="net.measurementlab.ndt7.android">
 
     <uses-permission android:name="android.permission.INTERNET" />
