+++ libndt7/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="net.measurementlab.ndt.android"/>
+<manifest package="net.measurementlab.ndt7.android"/>
